+++ utils/math.h
@@ -6,12 +6,12 @@
 #define M_E 2.71828182845904523536
 
 struct vec2_t {
-	float x;
-	float y;
+	int x;
+	int y;
 
 	vec2_t() : x(0), y(0) {} // Default constructor
 
-	vec2_t(float x, float y) : x(x), y(y) {}
+	vec2_t(int x, int y) : x(x), y(y) {}
 
 	vec2_t operator+(const vec2_t& other) {
 		return vec2_t(x + other.x, y + other.y);
@@ -21,11 +21,11 @@ struct vec2_t {
 		return vec2_t(x - other.x, y - other.y);
 	}
 
-	vec2_t operator*(const float& scalar) {
+	vec2_t operator*(const int& scalar) {
 		return vec2_t(x * scalar, y * scalar);
 	}
 
-	vec2_t operator/(const float& scalar) {
+	vec2_t operator/(const int& scalar) {
 		return vec2_t(x / scalar, y / scalar);
 	}
